src/app/components/milestone-list/milestone-list.component.spec.ts
import {
  beforeEach,
  beforeEachProviders,
  describe,
  expect,
  it,
  inject,
} from '@angular/core/testing';
import { ComponentFixture, TestComponentBuilder } from '@angular/compiler/testing';
import { Component } from '@angular/core';
import { By } from '@angular/platform-browser';
import { MilestoneListComponent } from './milestone-list.component';

describe('Component: MilestoneList', () => {
  let builder: TestComponentBuilder;

  beforeEachProviders(() => [MilestoneListComponent]);
  beforeEach(inject([TestComponentBuilder], function (tcb: TestComponentBuilder) {
    builder = tcb;
  }));

  it('should inject the component', inject([MilestoneListComponent],
      (component: MilestoneListComponent) => {
    expect(component).toBeTruthy();
  }));

  it('should create the component', inject([], () => {
    return builder.createAsync(MilestoneListComponentTestController)
      .then((fixture: ComponentFixture<any>) => {
        let query = fixture.debugElement.query(By.directive(MilestoneListComponent));
        expect(query).toBeTruthy();
        expect(query.componentInstance).toBeTruthy();
      });
  }));
});

@Component({
  selector: 'test',
  template: `
    <app-milestone-list></app-milestone-list>
  `,
  directives: [MilestoneListComponent]
})
class MilestoneListComponentTestController {
}

